requirements-dev.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
cfgv==3.4.0
discord.py==2.4.0
distlib==0.3.8
filelock==3.15.4
frozenlist==1.4.1
identify==2.6.0
idna==3.7
multidict==6.0.5
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
ruff==0.5.3
virtualenv==20.26.3
yarl==1.9.4
